app/requirements.txt
langchain_openai
langchain_experimental
langchain_community
langchain_core
langgraph
matplotlib
openai==1.56.1
pandas
plotly.express
psycopg2-binary
pydantic
scikit-learn
sqlalchemy
streamlit
streamlit-option-menu
typing

